requirements.txt
paho-mqtt
smbus
RPi.GPIO
